+++ src/test/java/runner/TestUtils.java
@@ -1,6 +1,8 @@
 package runner;
 
 import org.apache.commons.lang3.RandomStringUtils;
+import org.openqa.selenium.*;
+import org.openqa.selenium.support.ui.ExpectedCondition;
 
 public class TestUtils {
 
@@ -11,5 +13,40 @@ public class TestUtils {
 
     public static String getRandomStr() {
         return getRandomStr(15);
+    }
+
+    public static void scrollPageDown(WebDriver driver) {
+        ((JavascriptExecutor) driver).executeScript("window.scrollTo(0, document.body.scrollHeight)");
+    }
+
+    public static void scrollToElement(WebDriver driver, WebElement element) {
+        ((JavascriptExecutor) element).executeScript("arguments[0].scrollIntoView();", element);
+
+    }
+
+    public static ExpectedCondition<WebElement> waitElementStabilization(final By locator) {
+        return new ExpectedCondition<>() {
+            private WebElement element = null;
+            private Point location = null;
+
+            @Override
+            public WebElement apply(WebDriver driver) {
+                try {
+                    element = driver.findElement(locator);
+                } catch (NoSuchElementException e) {
+                    return null;
+                }
+
+                if (element.isDisplayed()) {
+                    Point location = element.getLocation();
+                    if (location.equals(this.location)) {
+                        return element;
+                    }
+                    this.location = location;
+                }
+
+                return null;
+            }
+        };
     }
 }
